adding ucmeanteacher to the framework
